requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

# Core PyTorch stack (CUDA 12.4)
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# Vision Transformer utilities
timm>=0.9.16
transformers>=4.39.0

# Optimizers
lion-pytorch>=0.1.0

# Miscellaneous
Pillow>=10.0
numpy>=1.23.0
